Changes for new token mapping
