--- redcap/data_source/pom.xml
@@ -1,37 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-  <name>MSKCC CMO Redcap Data Source</name>
-  <description>Redcap Source</description>
-  <artifactId>data_source</artifactId>
-  <version>0.1.0</version>
-  <groupId>org.mskcc.cmo.ks.redcap.source</groupId>
-
-  <parent>
-    <groupId>org.mskcc.cmo.ks.redcap</groupId>
-    <artifactId>redcap</artifactId>
-    <version>0.1.0</version>
-  </parent>
-
-  <dependencies>
-    <dependency>
-	  <groupId>org.springframework</groupId>
-	  <artifactId>spring-web</artifactId>
-    </dependency>
-    <dependency>
-	  <groupId>com.fasterxml.jackson.core</groupId>
-	  <artifactId>jackson-databind</artifactId>
-	</dependency>
-    <dependency>
-        <groupId>commons-lang</groupId>
-        <artifactId>commons-lang</artifactId>
-        <version>2.4</version>
-    </dependency>
-    <dependency>
-      <groupId>commons-cli</groupId>
-      <artifactId>commons-cli</artifactId>
-      <version>1.3</version>
-    </dependency>
-  </dependencies>
-</project>
-
