+++ redcap/redcap_pipeline/pom.xml
@@ -1,0 +1,47 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <name>MSKCC CMO Redcap Pipeline</name>
+  <description>Redcap Pipeline</description>
+  <artifactId>redcap_pipeline</artifactId>
+  <version>0.1.0</version>
+  <parent>
+    <groupId>org.mskcc.cmo.ks.redcap</groupId>
+    <artifactId>redcap</artifactId>
+    <version>0.1.0</version>
+  </parent>
+
+    <dependencies>
+    <dependency>
+        <groupId>org.springframework</groupId>
+        <artifactId>spring-web</artifactId>
+    </dependency>
+    <dependency>
+        <groupId>com.fasterxml.jackson.core</groupId>
+        <artifactId>jackson-databind</artifactId>
+      </dependency>
+    <dependency>
+        <groupId>commons-lang</groupId>
+        <artifactId>commons-lang</artifactId>
+        <version>2.4</version>
+    </dependency>
+    <dependency>
+      <groupId>commons-cli</groupId>
+      <artifactId>commons-cli</artifactId>
+      <version>1.3</version>
+    </dependency>    
+    <dependency>
+            <groupId>org.codehaus.jackson</groupId>
+            <artifactId>jackson-mapper-asl</artifactId>
+            <version>1.9.13</version>
+    </dependency>
+    <dependency>
+        <groupId>org.mskcc.cmo.ks.redcap.source</groupId>
+        <artifactId>redcap_source</artifactId>
+        <version>0.1.0</version>
+    </dependency>
+  </dependencies>
+    <properties>
+        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>1.8</maven.compiler.target>
+    </properties>
+</project>
