+++ redcap/redcap_pipeline/src/main/resources/application.properties.EXAMPLE
