+++ redcap/redcap_source/pom.xml
@@ -1,0 +1,27 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+  <modelVersion>4.0.0</modelVersion>
+  <name>MSKCC CMO Redcap Data Source</name>
+  <description>Redcap Source</description>
+  <artifactId>redcap_source</artifactId>
+  <version>0.1.0</version>
+  <groupId>org.mskcc.cmo.ks.redcap.source</groupId>
+
+  <parent>
+    <groupId>org.mskcc.cmo.ks.redcap</groupId>
+    <artifactId>redcap</artifactId>
+    <version>0.1.0</version>
+  </parent>
+
+  <dependencies>
+    <dependency>
+	  <groupId>org.springframework</groupId>
+	  <artifactId>spring-web</artifactId>
+    </dependency>
+    <dependency>
+	  <groupId>com.fasterxml.jackson.core</groupId>
+	  <artifactId>jackson-databind</artifactId>
+	</dependency>
+  </dependencies>
+</project>
+
